little bit of refacto
