+++ frontend/js/components/toast/toast.html
@@ -1,4 +1,4 @@
-<div class="toast" role="alert" aria-live="assertive" aria-atomic="true" data-bs-delay="30000">
+<div class="toast" role="alert" aria-live="assertive" aria-atomic="true" data-bs-delay="15000">
     <div class="toast-header">
         <img id="toast-img" width="25" height="25" src="https://png.pngtree.com/png-clipart/20190904/ourmid/pngtree-80-3d-text-png-image_18456.jpg" class="rounded me-2" alt="...">
         <strong id="toast-title" class="me-auto">Bootstrap</strong>
