+++ frontend/js/pages/home/utils.js
@@ -3,12 +3,7 @@ import { assembler } from '../../api/assembler.js';
 import { getUserAndDisplay, updateOtherFriendButton, updateStatusMsg } from './otherUserProfile.js';
 import { updateSocial, updateSocialFriendCard } from './social.js';
 import interactiveSocket from './socket.js';
-import { userTemplateComponent } from '../../components/userTemplate/userTemplate.js';
-import { fillOtherUserInfo } from './leftColumn.js';
-import { displayOtherUserProfile } from './otherUserProfile.js';
-import { displayToast } from './toastNotif.js';
-import { displayAlertMsg } from '../../utils/utilityFunctions.js';
-import { displayFriend } from './home.js';
+
 
 export async function handleFriendAction(actionObj) {
     const actionToMethod = {
@@ -56,87 +51,6 @@ export async function handleFriendAction(actionObj) {
     }
 }
 
-async function addNewUser(user) {
-    const everyoneContainer = document.getElementById('userDisplay');
-    const friendContainer = document.getElementById('friendDisplay');
-    const templateUser = await getTemplateUser(user);
-
-    if (!templateUser) return;
-
-    const seeProfileBtn = templateUser.querySelector('.card');
-    seeProfileBtn.addEventListener('click', displayOtherUserProfile)
-    appendToContainer(everyoneContainer, templateUser, user.id);
-    if (await checkIfFriend(user)) {
-        await changeFriendStatus(user.id, friendContainer, templateUser);
-    }
-}
-
-async function changeFriendStatus(userID, container, templateUser) {
-    const friendCard = container.querySelector(`div[data-id="${userID}"]`);
-
-    if (friendCard) {
-        updateBadgeColor(friendCard, container);
-    } else {
-        prependToContainer(container, templateUser.cloneNode(true), userID);
-    }
-}
-
-function updateBadgeColor(friendCard, container) {
-    const statusBadge = friendCard.querySelector('#badge');
-    if (statusBadge) {
-        statusBadge.style.backgroundColor = setStatus('ONL');
-        container.removeChild(friendCard);
-        container.insertBefore(friendCard, container.firstChild);
-    }
-}
-
-function appendToContainer(container, element, userID) {
-    if (!container.querySelector(`div[data-id="${userID}"]`)) {
-        element.setAttribute('data-id', userID);
-        container.appendChild(element);
-    } else {
-        console.log(`User with ID ${userID} already exists in the container.`);
-    }
-}
-
-function prependToContainer(container, element, userID) {
-    element.setAttribute('data-id', userID);
-    container.insertBefore(element, container.firstChild);
-}
-
-async function getTemplateUser(user) {
-    const templateUser = await userTemplateComponent();
-    if (!templateUser) {
-        console.error('Cannot find templateUser');
-        return null;
-    }
-    return fillOtherUserInfo(templateUser, user);
-}
-
-async function checkIfFriend(user) {
-    const response = await fetchFriendChange('GET', { id: user.id });
-    if (!response) return false;
-    
-    const friendStatus = await assembler(response);
-    return friendStatus && friendStatus.state === 'friend';
-}
-
-
-function updateOtherUsers(user) {
-    const userCards = document.querySelectorAll(`div[data-id="${user.id}"]`);
-    userCards.forEach(card => {
-        const avatarElement = card.querySelector('#user-avatar');
-        const nameElement = card.querySelector('#user-name');
-
-        if (avatarElement) {
-            avatarElement.src = user.avatar; // Update avatar image source
-            avatarElement.alt = user.nickname; // Update alt text
-        }
-        if (nameElement) {
-            nameElement.textContent = user.nickname; // Update user's name
-        }
-    });
-}
 
 export function setStatus(user) {
     switch (user) {
@@ -150,6 +64,7 @@ export function setStatus(user) {
             return 'gray';
     }
 }
+
 
 export function getMyID() {
     let userID = sessionStorage.getItem('user_id');
@@ -171,122 +86,17 @@ export function getMyID() {
     return userID;
 }
 
-//////////////////////////// SOCKET FUNCTIONS ////////////////////////////
 
-// To use when user update his profile (avatar/nickname)
-export async function updateSpecificUser(userID) {
-    console.log("UPDATE USER")
-    const apiParam = { id: userID };
-    const response = await fetchUser('GET', apiParam);
+export async function fetchUserById(userID) {
+    const response = await fetchUser('GET', { id: userID });
     if (!response)
         return;
     const assemble = await assembler(response);
     if (typeof assemble !== 'object' || assemble === null) {
-        console.log(assemble);
         return;
-    } else {
-        updateOtherUsers(assemble[0]);
     }
+    return assemble[0];
 }
 
-// To use when user login
-export async function newUser(userID) {
-    console.log("NEW USER")
-    const apiParam = { id: userID };
-    if (getMyID() == userID) 
-        return;
-    try {
-        const response = await fetchUser('GET', apiParam);
-        if (!response)
-            return;
-        
-        const assemble = await assembler(response);
-        if (typeof assemble !== 'object' || assemble === null) {
-            console.log(assemble);
-            return;
-        }
-        addNewUser(assemble[0]);
-    } catch (error) {
-        console.error('Error in newUser:', error);
-    }
-    console.log("END NEW USER")
-}
 
-// To use when user logout
-export async function removeUser(userID) {
-    console.log("REMOVE USER")
-    const everyoneContainer = document.getElementById('userDisplay');
-    const friendContainer = document.getElementById('friendDisplay');
-
-    const userCards = document.querySelectorAll(`div[data-id="${userID}"]`);
-    userCards.forEach(card => {
-        if (everyoneContainer.contains(card)) {
-            everyoneContainer.removeChild(card);
-        } else if (friendContainer.contains(card)) {
-            const statusBadge = card.querySelector('#badge');
-            if (statusBadge) {
-                statusBadge.style.backgroundColor = setStatus('OFF');
-                friendContainer.removeChild(card);
-                friendContainer.appendChild(card);
-            }
-        }
-    });
-    console.log("END REMOVE USER")
-}
-
-function updateModalIfOpen() {
-    const otherUserModal = document.getElementById('otherUserInfo')
-    console.log(otherUserModal)
-    if (otherUserModal.classList.contains('show'))  { // ERROR HERE
-        const content = otherUserModal.querySelector('.modal-content');
-        if (content.dataset.id)
-            getUserAndDisplay(content.dataset.id);
-    }
-}
-
-export function handleSocialUpdate(rType, currentUser, otherUserId) {
-    const userId = getMyID();
-    console.log("START SOCIAL UPDATE")
-    if (!userId || userId == currentUser || userId == otherUserId) {
-        console.log("IN IF")
-        updateSocial();
-        if (userId == otherUserId)
-            updateModalIfOpen()
-        displayFriend();
-        createNotifications(rType, userId, currentUser, otherUserId);
-    }
-
-    console.log("END SOCIAL UPDATE");
-}
-
-function createNotifications(rType, userId, otherUserId, currentUser) {
-    let toastMsg = "";
-    let toastTitle = "";
-    let imgUrl = "https://www.shutterstock.com/image-vector/friends-request-icon-isolated-sign-260nw-1591730662.jpg";
-
-    if (userId == currentUser) {
-        switch (rType) {
-            case "add":
-                toastMsg = "You have received a friend request!";
-                toastTitle = "Friend Request";
-                break;
-            case "accept":
-                toastMsg = "Your friend request was accepted!";
-                toastTitle = "Request Accepted";
-                break;
-            case "refuse":
-                toastMsg = "Your friend request has been refused.";
-                toastTitle = "Request Refused";
-                break;
-            case "unfriend":
-                toastMsg = "You have been unfriended.";
-                toastTitle = "Unfriended";
-                break;
-            default:
-                return;
-        }
-
-        if (toastMsg)
-            displayToast(toastMsg, toastTitle, imgUrl);
-    }
-}
+//////////////////////////// SOCKET FUNCTIONS ////////////////////////////
